task: #3555534 Since symfony/validator 7.4: Support for evaluating options in the base Constraint class is deprecated. Initialize properties in the constructor instead.
By: longwave
By: godotislate
By: larowlan
(cherry picked from commit b6627d2bc3aefde033d1c7a90922839a27d04470)

+++ core/modules/locale/src/Form/ImportForm.php
@@ -100,7 +100,7 @@ class ImportForm extends FormBase {
 
     $validators = [
       'FileExtension' => ['extensions' => 'po'],
-      'FileSizeLimit' => ['fileLimit' => Environment::getUploadMaxSize()],
+      'FileSizeLimit' => ['fileLimit' => (int) Environment::getUploadMaxSize()],
       'FileEncoding' => ['encodings' => ['UTF-8']],
     ];
     $form['file'] = [
